Comments, requirement.txt added, tool name changed. Test dir comming soon.
